refactor: Operator 파일 분리 #2
